move logic into useRunSeerAnalysis.tsx
